Write encryptor (#11)
* [+] Encrypt write payloads

* Changed db table column types

* Refactor

* fix: Pipfile.lock actualization

* refactor

refactor: Payload generator traceback printing

style: Reformat with keeping indents on empty lines

fix: Postgres connector repr

refactor: Reading files within context manager

style: Comment why line commented (lol)

refactor: Encryptor types fix and documentation

fix: usage of Encryptor methods

style: Variable rename, removed print

* refactor: phe, ope ciphers in base Encryptor

* refactor: Write tokens encryptor refactor

* feature: Decrypt string method

* style: Reformat

* style: Type annotations

* refactor: Variable rename

* refactor: Variable rename

+++ docker_images/timescale/002_create_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE laptop_meas (
     time TIMESTAMPTZ NOT NULL,
     hostname VARCHAR(256),
-    cpu_percent DOUBLE PRECISION,
-    cpu_freq DOUBLE PRECISION,
-    memory_used BIGINT
+    cpu_percent NUMERIC,
+    cpu_freq NUMERIC,
+    memory_used NUMERIC
 )
